make LadderGame.swift file
